separated out receipt taking page, created a home page, utilized reactnavigation
