+++ package 2.json
@@ -8,34 +8,17 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@expo-google-fonts/dev": "^0.2.0",
-    "@expo-google-fonts/inter": "^0.2.0",
-    "@expo-google-fonts/lato": "^0.2.0",
     "@google-cloud/vision": "^2.3.8",
-    "@react-navigation/bottom-tabs": "^6.0.5",
-    "@react-navigation/material-bottom-tabs": "^6.0.5",
-    "@react-navigation/material-top-tabs": "^6.0.2",
     "@react-navigation/native": "^6.0.2",
-    "@react-navigation/native-stack": "^6.1.0",
-    "@react-navigation/stack": "^6.0.7",
     "expo": "~42.0.1",
-    "expo-app-loading": "^1.1.2",
     "expo-camera": "~11.2.2",
     "expo-constants": "~11.0.1",
-    "expo-font": "~9.2.1",
     "expo-image-picker": "^10.2.2",
     "expo-status-bar": "~1.0.4",
     "firebase": "^8.2.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
-    "react-native-gesture-handler": "^1.10.3",
-    "react-native-pager-view": "^5.4.0",
-    "react-native-paper": "^4.9.2",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
-    "react-native-tab-view": "^3.1.1",
-    "react-native-vector-icons": "^8.1.0",
     "react-native-web": "~0.13.12",
     "uuid": "^3.4.0"
   },
